# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/request-otp": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/admin/password-challenge": {
      "post": {
        "operationId": "AuthController_passwordChallenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPasswordChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/admin/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyAdminOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOtpVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyAuthOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyAuthOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/auth/onboarding/finalize": {
      "post": {
        "operationId": "AuthController_finalizeOnboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
...
    "/api/v1/admin/admins": {
      "get": {
        "operationId": "AdminIdentityController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminIdentity"]
      },
      "post": {
        "operationId": "AdminIdentityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminIdentity"]
      }
    },
...
    "/api/v1/admin/admins/{adminId}": {
      "patch": {
        "operationId": "AdminIdentityController_update",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminIdentity"]
      }
    },
    "/api/v1/admin/admins/{adminId}/archive": {
      "post": {
        "operationId": "AdminIdentityController_archive",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminIdentity"]
      }
    },
    "/api/v1/admin/admins/{adminId}/unarchive": {
      "post": {
        "operationId": "AdminIdentityController_unarchive",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminIdentity"]
      }
    },
...
  },
  "info": {
    "title": "School Transport API",
    "description": "Canonical REST contract for the school transport MVP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {}
      },
      "AdminPasswordChallengeDto": {
        "type": "object",
        "properties": {}
      },
      "AdminOtpVerificationDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyAuthOtpDto": {
        "type": "object",
        "properties": {}
      },
      "FinalizeOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAdminDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}